+++ tests/RouterTest.php
@@ -9,6 +9,8 @@
 namespace Slim\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Slim\Dispatcher;
+use Slim\RoutingResults;
 use Slim\Router;
 
 class RouterTest extends TestCase
@@ -205,7 +207,7 @@ class RouterTest extends TestCase
         $class = new \ReflectionClass($this->router);
         $method = $class->getMethod('createDispatcher');
         $method->setAccessible(true);
-        $this->assertInstanceOf('\FastRoute\Dispatcher', $method->invoke($this->router));
+        $this->assertInstanceOf(Dispatcher::class, $method->invoke($this->router));
     }
 
     public function testSetDispatcher()
@@ -213,11 +215,11 @@ class RouterTest extends TestCase
         $this->router->setDispatcher(\FastRoute\simpleDispatcher(function ($r) {
             $r->addRoute('GET', '/', function () {
             });
-        }));
+        }, ['dispatcher' => Dispatcher::class]));
         $class = new \ReflectionClass($this->router);
         $prop = $class->getProperty('dispatcher');
         $prop->setAccessible(true);
-        $this->assertInstanceOf('\FastRoute\Dispatcher', $prop->getValue($this->router));
+        $this->assertInstanceOf(Dispatcher::class, $prop->getValue($this->router));
     }
 
     /**
@@ -362,7 +364,7 @@ class RouterTest extends TestCase
         $method->setAccessible(true);
 
         $dispatcher = $method->invoke($this->router);
-        $this->assertInstanceOf('\FastRoute\Dispatcher', $dispatcher);
+        $this->assertInstanceOf(Dispatcher::class, $dispatcher);
         $this->assertFileExists($cacheFile, 'cache file was not created');
 
         // instantiate a new router & load the cached routes file & see if
@@ -375,8 +377,12 @@ class RouterTest extends TestCase
         $method->setAccessible(true);
 
         $dispatcher2 = $method->invoke($this->router);
+
+        /**
+         * @var RoutingResults $result
+         */
         $result = $dispatcher2->dispatch('GET', '/hello/josh/lockhart');
-        $this->assertSame(\FastRoute\Dispatcher::FOUND, $result[0]);
+        $this->assertSame(Dispatcher::FOUND, $result->getRouteStatus());
 
         unlink($cacheFile);
     }
@@ -415,7 +421,7 @@ class RouterTest extends TestCase
         $queryParams = ['a' => 'b', 'c' => 'd'];
 
         //create a router that mocks the pathFor with expected args
-        $router = $this->getMockBuilder('\Slim\Router')->setMethods(['pathFor'])->getMock();
+        $router = $this->getMockBuilder(Router::class)->setMethods(['pathFor'])->getMock();
         $router->expects($this->once())->method('pathFor')->with($name, $data, $queryParams);
         $router->urlFor($name, $data, $queryParams);